chore(test): drop ignored live-agent e2e and L1 leftovers
- Delete real Codex/Docker/ACP probe cases CI never ran
- Keep offline fail-closed CLI checks in the default suite
- python-core now ignores only tests/registry
